refactor: group confirmationdialog
