+++ frontend/src/components/PersonModal/index.tsx
@@ -1,6 +1,6 @@
 import { X } from "lucide-react";
 import { useCallback, useRef, useState } from "react";
-import ConfirmationDialog from "./ConfirmationDialog";
+import ConfirmationDialog from "../ConfirmationDialog";
 import PersonForm from "./PersonForm";
 import "./PersonModal.css";
